requirements.txt
dataclasses-json==0.6.7
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.24
loguru==0.7.3
numpy==2.2.4
openai==1.70.0
pandas==2.2.3
pillow==11.1.0
pydantic==2.11.2
pypdf==5.4.0
python-dotenv==1.1.0
PyYAML==6.0.2
qdrant-client==1.13.3
requests==2.32.3
SQLAlchemy==2.0.40
streamlit==1.44.1
tenacity==9.1.2
tqdm==4.67.1
urllib3==2.3.0